Fix some compile-time warnings

+++ integration/non_event_data/src/MuonGeometry.cpp
@@ -20,7 +20,7 @@ Consumers::MuonGeometry::MuonGeometry(
 void Consumers::MuonGeometry::consume(std::vector<char> const& data)
 {
   const char* raw_input = data.data();
-  for (int i = 0; i < n_preamble_blocks; i++) {
+  for (size_t i = 0; i < n_preamble_blocks; i++) {
     size_t size;
     std::copy_n((size_t*) raw_input, 1, &size);
     raw_input += sizeof(size_t);
